Refactor payment processing and order management: Update IPaymentService to accept amount, modify OrderDTO and ActiveOrder for Long types, enhance error handling in ReservationService, and improve order repository methods.

+++ src/main/java/ticketsystem/DTO/OrderDTO.java
@@ -9,14 +9,14 @@ public class OrderDTO {
     private String eventName;
     private String location;
     private long memberId;
-    private int companyId;
+    private Long companyId;
 
     //for json 
     public OrderDTO() {
     }
 
     public OrderDTO(int purchaseId, List<TicketDTO> tickets, String eventName, 
-                    String location, long memberId, int companyId) {
+                    String location, long memberId, Long companyId) {
         this.purchaseId = purchaseId;
         this.tickets = tickets;
         this.eventName = eventName;
@@ -65,11 +65,11 @@ public class OrderDTO {
         this.memberId = memberId;
     }
 
-    public int getCompanyId() {
+    public Long getCompanyId() {
         return companyId;
     }
 
-    public void setCompanyId(int companyId) {
+    public void setCompanyId(Long companyId) {
         this.companyId = companyId;
     }
 }
